fix: use double pointer to implement setDefaultValue

+++ internal/service/plugin_decoder.go
@@ -50,7 +50,7 @@ func UploadPluginPkg(
 		return exception.BadRequestError(errors.Join(err, errors.New("failed to save package"))).ToResponse()
 	}
 
-	if config.ForceVerifyingSignature || verify_signature {
+	if config.ForceVerifyingSignature != nil && *config.ForceVerifyingSignature || verify_signature {
 		if !declaration.Verified {
 			return exception.BadRequestError(errors.Join(err, errors.New(
 				"plugin verification has been enabled, and the plugin you want to install has a bad signature",
@@ -138,7 +138,7 @@ func UploadPluginBundle(
 						return exception.InternalServerError(errors.Join(errors.New("failed to save package"), err)).ToResponse()
 					}
 
-					if config.ForceVerifyingSignature || verify_signature {
+					if config.ForceVerifyingSignature != nil && *config.ForceVerifyingSignature || verify_signature {
 						if !declaration.Verified {
 							return exception.BadRequestError(errors.Join(errors.New(
 								"plugin verification has been enabled, and the plugin you want to install has a bad signature",
